+++ core/graph/__init__.py
@@ -1,6 +1,8 @@
+from .export import export_graphml, export_json_ld
 from .migrate import migrate_graph_data, migrate_json_file
 from .pruning import PruningPolicy, apply_pruning_policy
 from .rules_engine import RulesEngineCycleResult
+from .snapshots import GraphSnapshotManager
 from .sqlite_backend import SQLiteGraphBackend
 from .universal_living_graph import UniversalLivingGraph
 
@@ -12,4 +14,7 @@ __all__ = [
     "apply_pruning_policy",
     "migrate_graph_data",
     "migrate_json_file",
+    "GraphSnapshotManager",
+    "export_graphml",
+    "export_json_ld",
 ]